Rename simulation test to crucible
