+++ packages/cli/test/scripts/e2e_test_env.ts
@@ -1,9 +1,9 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import path from "node:path";
-import {BeaconClient, ExecutionClient} from "../utils/simulation/interfaces.js";
-import {SimulationEnvironment} from "../utils/simulation/simulationEnvironment.js";
-import {defineSimTestConfig, logFilesDir} from "../utils/simulation/utils/index.js";
-import {connectAllNodes} from "../utils/simulation/utils/network.js";
+import {BeaconClient, ExecutionClient} from "../utils/crucible/interfaces.js";
+import {SimulationEnvironment} from "../utils/crucible/simulationEnvironment.js";
+import {defineSimTestConfig, logFilesDir} from "../utils/crucible/utils/index.js";
+import {connectAllNodes} from "../utils/crucible/utils/network.js";
 
 const altairForkEpoch = 1;
 const bellatrixForkEpoch = 2;
