+++ packages/cli/test/sim/backupEthProvider.test.ts
@@ -1,11 +1,11 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import path from "node:path";
 import {activePreset} from "@lodestar/params";
-import {SimulationEnvironment} from "../utils/simulation/simulationEnvironment.js";
-import {nodeAssertion} from "../utils/simulation/assertions/nodeAssertion.js";
-import {AssertionMatch, BeaconClient, ExecutionClient} from "../utils/simulation/interfaces.js";
-import {defineSimTestConfig, logFilesDir, replaceIpFromUrl} from "../utils/simulation/utils/index.js";
-import {connectAllNodes, waitForSlot} from "../utils/simulation/utils/network.js";
+import {SimulationEnvironment} from "../utils/crucible/simulationEnvironment.js";
+import {nodeAssertion} from "../utils/crucible/assertions/nodeAssertion.js";
+import {AssertionMatch, BeaconClient, ExecutionClient} from "../utils/crucible/interfaces.js";
+import {defineSimTestConfig, logFilesDir, replaceIpFromUrl} from "../utils/crucible/utils/index.js";
+import {connectAllNodes, waitForSlot} from "../utils/crucible/utils/network.js";
 
 const altairForkEpoch = 2;
 const bellatrixForkEpoch = 4;
