+++ packages/cli/test/utils/crucible/assertions/defaults/attestationParticipationAssertion.ts
