+++ packages/cli/test/utils/crucible/assertions/defaults/finalizedAssertion.ts
