+++ packages/cli/test/utils/crucible/assertions/defaults/headAssertion.ts
