+++ packages/cli/test/utils/crucible/assertions/defaults/syncCommitteeParticipationAssertion.ts
