+++ packages/cli/test/utils/crucible/assertions/executionHeadAssertion.ts
