+++ packages/cli/test/utils/crucible/clients/beacon/index.ts
