+++ packages/cli/test/utils/crucible/clients/beacon/lighthouse.ts
