+++ packages/cli/test/utils/crucible/clients/beacon/lodestar.ts
