+++ packages/cli/test/utils/crucible/clients/validator/lighthouse.ts
