+++ packages/cli/test/utils/crucible/clients/validator/lodestar.ts
