+++ packages/cli/test/utils/crucible/simulationEnvironment.ts
