+++ packages/cli/test/utils/crucible/simulationTracker.ts
