+++ packages/cli/test/utils/crucible/tableRenderer.ts
